requirements.txt
beautifulsoup4==4.10.0
certifi==2021.10.8
charset-normalizer==2.0.12
idna==3.3
numpy==1.22.3
pandas==1.4.1
python-dateutil==2.8.2
pytz==2022.1
requests==2.27.1
six==1.16.0
soupsieve==2.3.1
urllib3==1.26.9
SQLAlchemy~=1.4.35
asttokens==2.0.5
backcall==0.2.0
beautifulsoup4==4.10.0
certifi==2021.10.8
charset-normalizer==2.0.12
cycler==0.11.0
decorator==5.1.1
ecos==2.0.10
et-xmlfile==1.1.0
executing==0.8.3
fonttools==4.33.3
idna==3.3
ipython==8.2.0
jedi==0.18.1
joblib==1.1.0
kaleido==0.2.1
kiwisolver==1.4.2
lxml==4.8.0
matplotlib==3.5.2
matplotlib-inline==0.1.3
numexpr==2.8.1
numpy==1.22.3
openpyxl==3.0.9
osqp==0.6.2.post5
packaging==21.3
pandas==1.4.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.0
plotly==5.8.0
prompt-toolkit==3.0.29
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.11.2
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
qdldl==0.1.5.post2
requests==2.27.1
scikit-learn==1.0.2
scikit-survival==0.17.2
scipy==1.8.0
seaborn==0.11.2
six==1.16.0
soupsieve==2.3.1
stack-data==0.2.0
stats==0.1.2a0
tenacity==8.0.1
threadpoolctl==3.1.0
traitlets==5.1.1
urllib3==1.26.9
wcwidth==0.2.5
